+++ src/legacy/bench_mem_dicuq.zig
@@ -10,7 +10,7 @@ const F = buildconfig.F;
 const DEFAULT_OUT_DIR = "out/bench_stats_mem_dicuq";
 const DEFAULT_DATA_DIR = "data/FE/platehole3d_6mr_63f/";
 const DEFAULT_PIXELS_NUM = [2]u32{ 2464, 2056 };
-const DEFAULT_SUB_SAMPLE: u8 = 2;
+const DEFAULT_SUB_SAMPLE: u16 = 2;
 const DEFAULT_FOCAL_LENG: F = @floatCast(50.0e-3);
 const DEFAULT_PIXELS_SIZE = [2]F{
     @floatCast(3.45e-6),
